+++ components/goals-tasks-nav-tree.tsx
@@ -207,9 +207,8 @@ const GoalsTaskNavTree = ({ projectId }: GoalsTaskNavTreeProps) => {
   return (
     <>
       <div className="flex flex-col items-center">
-        <Label.Root>Goals & Tasks Explorer</Label.Root>
+        <Label.Root className="font-semibold" style={{ marginBottom: '0px' }}>Goals & Tasks Explorer</Label.Root>
       </div>
-      <br />
       <Tree data={goalsTasksTreeData}>{renderGoalOrTaskNode}</Tree>
     </>
   )